Mensaje para cuando no hay ningún recurso

+++ views/resources/view_all.php
@@ -23,21 +23,29 @@ while ($fila = $result->fetch_object()) {
 }
 */
 
-echo "<table border=1>";
-foreach ($data['resources'] as $resources) {
 
-    echo "<tr>";
-    echo "<td>" . $resources['name'] . "</td>";
-    echo "<td>" . $resources['description'] . "</td>";
-    echo "<td>" . $resources['location'] . "</td>";
-    echo "<td>" . "<img src='" . $resources['image'] . "' width='50px'/>" . "</td>";
+if(empty($data['resources']) != 1){ // Comprobamos si hay datos
+    echo "<table border=1>";
+    foreach ($data['resources'] as $resources) {
 
-    echo "<td><a class='btn btn-outline-info' href='index.php?action=modify_resource_form&id=" . $resources['id'] . "'>Modificar</a></td>";
-    echo "<td><a class='btn btn-outline-danger confirmacion' href='index.php?controller=ResourcesController&action=deleteResource&id=" . $resources['id'] . "'>Borrar</a></td>";
-    echo "</tr>";
+        echo "<tr>";
+        echo "<td>" . $resources['name'] . "</td>";
+        echo "<td>" . $resources['description'] . "</td>";
+        echo "<td>" . $resources['location'] . "</td>";
+        echo "<td>" . "<img src='" . $resources['image'] . "' width='50px'/>" . "</td>";
+
+        echo "<td><a class='btn btn-outline-info' href='index.php?action=modify_resource_form&id=" . $resources['id'] . "'>Modificar</a></td>";
+        echo "<td><a class='btn btn-outline-danger confirmacion' href='index.php?controller=ResourcesController&action=deleteResource&id=" . $resources['id'] . "'>Borrar</a></td>";
+        echo "</tr>";
+    }
+
+    echo "</table>";
+}else{
+    echo "No se han encontrado recursos ❗";
 }
 
-echo "</table>";
+
+echo "<br/><br/>";
 
 /* Borrar
 foreach ($data['permissions'] as $permission) {
